Create RdbmsConnection from DBConnection
