+++ packages/shopify-app-session-storage/src/index.ts
@@ -2,3 +2,4 @@ export * from './session-storage';
 export * from './session-storage-migration';
 export * from './rdbms-session-storage-migrator';
 export * from './abstract-migration-engine';
+export * from './db-connection';
